Compact client/service provider parser
Use just one ListCommandParser, using the ContactType enum to ensure
correct behaviour
