+++ src/test/data/Account/accountlist.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$xJlS90cjUkRwZ1h1w5XcZUxVdclXV5u1GXBjQnXPZLg</password>
+        <password>$31$16$EzsQPooNo8j_RoQ1NDLNAMbIQPol0IrdVRHcvWIAa24</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
