+++ src/test/data/RegisterAccountCommandTest/accountlist_withrootonly.xml
@@ -2,12 +2,12 @@
 <accountmanager>
     <accounts>
         <username>rootUser</username>
-        <password>$31$16$4nTBeIWH2wHpYXGbyy4EX_r_hJ8Cx0NQDZ-2wNsjssE</password>
+        <password>$31$16$3fbRbkNYPQtM_8wftKwvn5-zHKzzsC1sC2yaoZdlaVM</password>
         <role>SUPER_USER</role>
     </accounts>
     <accounts>
         <username>username123</username>
-        <password>$31$16$gSv0IHBWtPvTLuwELQ_2Bx4KGjae6cb3Gui3WZDz-sI</password>
+        <password>$31$16$zqiUF79tNBIKqAoiWXqFiNGoS75_83kMBz69bte-HGA</password>
         <role>SUPER_USER</role>
     </accounts>
 </accountmanager>
