+++ src/test/data/XmlAccountStorageTest/accountlistToChangePassword.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$7BirkXymljTjRnasZkwbQDVxGjJTDCO6gqMWecDQZHI</password>
+        <password>$31$16$oSJKNWsb77a-xvUeaBNX1WgS7dHGBJLVRbgW1N-lKlA</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
