+++ src/test/data/XmlUtilTest/accountlist.xml
@@ -12,7 +12,7 @@
     </accounts>
     <accounts>
         <username>whiterose</username>
-        <password>$31$16$lQRytOr6jb_hXz53BPZu7vrrRsdhNmn0BhtE7CTnpgY</password>
+        <password>$31$16$UDgJjc-x6B_OzgLTcT936negMRaSYgpWTYOT6cC7BiQ</password>
         <role>READ_ONLY_USER</role>
     </accounts>
 </accountmanager>
