+++ src/test/java/seedu/address/logic/commands/ListCommandTest.java
@@ -6,7 +6,6 @@ import static org.junit.Assert.assertTrue;
 import static seedu.address.commons.core.Messages.MESSAGE_LIST_ALL_PERSON;
 import static seedu.address.commons.core.Messages.MESSAGE_PERSONS_LISTED_OVERVIEW;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.logic.parser.ListClientCommandParser.CONTACT_FILTER_CLIENT;
 import static seedu.address.testutil.TypicalContacts.ALICE;
 import static seedu.address.testutil.TypicalContacts.BENSON;
 import static seedu.address.testutil.TypicalContacts.CARL;
@@ -46,14 +45,14 @@ public class ListCommandTest {
         ContactContainsKeywordsPredicate secondPredicate =
                 new ContactContainsKeywordsPredicate(new ClientBuilder().withName("second").build());
 
-        ListCommand findFirstCommand = new ListCommand(firstPredicate, CONTACT_FILTER_CLIENT);
-        ListCommand findSecondCommand = new ListCommand(secondPredicate, CONTACT_FILTER_CLIENT);
+        ListCommand findFirstCommand = new ListCommand(firstPredicate, ContactType.CLIENT.getFilter());
+        ListCommand findSecondCommand = new ListCommand(secondPredicate, ContactType.CLIENT.getFilter());
 
         // same object -> returns true
         assertTrue(findFirstCommand.equals(findFirstCommand));
 
         // same values -> returns true
-        ListCommand findFirstCommandCopy = new ListCommand(firstPredicate, CONTACT_FILTER_CLIENT);
+        ListCommand findFirstCommandCopy = new ListCommand(firstPredicate, ContactType.CLIENT.getFilter());
         assertTrue(findFirstCommand.equals(findFirstCommandCopy));
 
         // different types -> returns false
@@ -70,7 +69,7 @@ public class ListCommandTest {
     public void execute_zeroKeywords_everyPersonFound() {
         String expectedMessage = MESSAGE_LIST_ALL_PERSON;
         ContactContainsKeywordsPredicate predicate = preparePredicate();
-        ListCommand command = new ListCommand(predicate, CONTACT_FILTER_CLIENT);
+        ListCommand command = new ListCommand(predicate, ContactType.CLIENT.getFilter());
         expectedModel.updateFilteredContactList(predicate.and(ContactType.CLIENT.getFilter()));
         assertCommandSuccess(command, model, commandHistory, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(ALICE, BENSON, CARL, DANIEL, ELLE, FIONA, GEORGE), model.getFilteredContactList());
@@ -82,7 +81,7 @@ public class ListCommandTest {
         /* Case: List with 3 person's name -> 0 persons found */
         String expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 0);
         ContactContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        ListCommand command = new ListCommand(predicate, CONTACT_FILTER_CLIENT);
+        ListCommand command = new ListCommand(predicate, ContactType.CLIENT.getFilter());
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, commandHistory, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(), model.getFilteredContactList());
@@ -90,7 +89,7 @@ public class ListCommandTest {
         /* Case: List with 1 person's name -> 1 persons found */
         expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 1);
         predicate = preparePredicate("Benson Meier");
-        command = new ListCommand(predicate, CONTACT_FILTER_CLIENT);
+        command = new ListCommand(predicate, ContactType.CLIENT.getFilter());
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, commandHistory, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(BENSON), model.getFilteredContactList());
